+++ database/migrations/2019_09_04_170511_create_answers_table.php
@@ -15,8 +15,8 @@ class CreateAnswersTable extends Migration
     {
         Schema::create('answers', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('qustion_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('qustion_id');
+            $table->unsignedBigInteger('user_id');
             $table->text('body');
             $table->integer('votes_count')->default(0);
             $table->timestamps();
